@@ -9,24 +9,31 @@
     "license": "MIT",
     "require": {
         "php": "^8.1",
+        "ext-zip": "*",
         "ariaieboy/filament-jalali-datetimepicker": "^2.0",
         "awcodes/filament-gravatar": "^2.0",
         "bezhansalleh/filament-shield": "^3.0",
         "filament/filament": "^3.0-stable",
         "filament/infolists": "^3.0-stable",
         "filament/notifications": "^3.1",
+        "filament/spatie-laravel-media-library-plugin": "^3.2",
         "filament/spatie-laravel-settings-plugin": "^3.1",
         "guzzlehttp/guzzle": "^7.2",
         "hasnayeen/themes": "^3.0",
         "hekmatinasser/verta": "^8.3",
-        "laravel/framework": "^10.10",
+        "jaocero/activity-timeline": "^1.2",
+        "laravel/framework": "^11.0",
         "laravel/horizon": "^5.22",
-        "laravel/octane": "^2.1",
+        "laravel/octane": "^2.3",
         "laravel/pulse": "^1.0@beta",
-        "laravel/sanctum": "^3.3",
+        "laravel/reverb": "@beta",
+        "laravel/sanctum": "^4.0",
         "laravel/tinker": "^2.8",
+        "leandrocfe/filament-apex-charts": "^3.1",
         "phpsa/filament-password-reveal": "^2.0",
-        "pusher/pusher-php-server": "^7.2",
+        "shuvroroy/filament-spatie-laravel-backup": "^2.1",
+        "spatie/laravel-activitylog": "^4.8",
+        "spatie/laravel-backup": "^8.8",
         "spatie/laravel-settings": "^3.2"
     },
     "require-dev": {
@@ -34,7 +41,7 @@
         "laravel/pint": "^1.13",
         "laravel/sail": "^1.18",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^7.0",
+        "nunomaduro/collision": "^8.1",
         "phpunit/phpunit": "^10.1",
         "spatie/laravel-ignition": "^2.0"
     },
