@@ -3,10 +3,11 @@
 namespace App\Filament\Resources\ProjectResource\Actions;
 
 use App\Models\Test;
-use App\Settings\GeneralSettings;
+use Filament\Forms\Components\TextInput;
+use Filament\Notifications\Notification;
 use Filament\Tables\Actions\Action;
 use Illuminate\Support\Facades\Auth;
-use Filament\Forms\Components\Select;
+use Illuminate\Support\Facades\DB;
 
 class RenewalAction extends Action
 {
@@ -15,23 +16,56 @@ class RenewalAction extends Action
 
         parent::setUp();
         $this->label('تمدید')
-            ->outlined()
-            ->form(function () {
-                $renewalRangeDuration = (int)app(GeneralSettings::class)->renewalDurationTime;
-                $renewalOptions = [];
-                for ($i = $renewalRangeDuration; $i <= 120; $i += $renewalRangeDuration) {
-                    $renewalOptions[$i] = $i . ' دقیقه ';
+            ->button()
+            ->form(
+                fn () => [
+                    TextInput::make('body')->label('متن')->required()->maxLength(100),
+                ]
+            )
+            ->action(function (Test $record, array $data) {
+
+                if ($record->projectTest->project->isFinished()) {
+                    Notification::make()
+                        ->title('.قادر  به انجام این عملیات نیستید')
+                        ->danger()
+                        ->send();
+                } else {
+                    if (!$record->projectTest->isAbleToRenewal()) {
+                        if (Auth::user()->can('renewal_project_test_project')) {
+                            $this->doRenewal($record, $data, true);
+                        } else {
+                            Notification::make()
+                                ->title('.قادر  به انجام این عملیات نیستید')
+                                ->danger()
+                                ->send();
+                        }
+                    } else {
+                        $this->doRenewal($record, $data);
+                    }
                 }
-                return[
-                    Select::make('renewal_duration')->label('مدت تمدید')->options($renewalOptions)->native(false)
-                ];
             })
-            ->action(fn (Test $record, array $data) => $record->projectTest->renewal($data['renewal_duration']))
-            ->requiresConfirmation()
-            ->hidden(
-                fn (Test $record): bool =>
-                !Auth::user()->can('renewal_project_test_project')
-                || !$record->projectTest->isAbleToRenewal()
-            );
+            ->requiresConfirmation();
+    }
+
+    private function doRenewal(Test $test, array $data, bool $force = false): void
+    {
+        DB::transaction(function () use ($test, $data) {
+            $test->projectTest->renewal();
+        });
+        activity()
+            ->event('renewal')
+            ->useLog('projects')
+            ->performedOn($test->projectTest->project)
+            ->causedBy(Auth::user())
+            ->log(" آزمایش  $test->title "
+                ." توسط "
+                .$test->projectTest->project->user->name
+      .  " تمدید شد "
+            .($force ? "و باعث افزایش زمان پروژه شد." : "."));
+        Notification::make()
+            ->title('تمدید با موفقیت انجام شد')
+            ->success()
+            ->send();
+        redirect("/admin/projects/{$test->projectTest->project->id}");
     }
 }
