@@ -2,11 +2,21 @@
 
 namespace App\Filament\Resources;
 
-use App\Models\Test;
+use App\Filament\Resources\ProjectResource\Actions\Bulk\ContinueAction;
+use App\Filament\Resources\ProjectResource\Actions\Bulk\PauseAction;
+use App\Filament\Resources\ProjectResource\RelationManagers\NotesRelationManager;
+use App\Models\Note;
 use App\Models\Project;
+use App\Models\User;
+use App\Tables\Columns\NewCountDownColumn;
+use Ariaieboy\FilamentJalaliDatetimepicker\Forms\Components\JalaliDatePicker;
+use Filament\Forms\Components\TextInput;
 use Filament\Forms\Get;
 use App\Models\Product;
 use Filament\Forms\Form;
+use Filament\Tables\Actions\BulkActionGroup;
+use Filament\Tables\Filters\Filter;
+use Filament\Tables\Filters\SelectFilter;
 use Filament\Tables\Table;
 use Filament\Resources\Resource;
 use Filament\Infolists\Infolist;
@@ -15,7 +25,6 @@ use Filament\Forms\Components\Select;
 use Filament\Tables\Columns\TextColumn;
 use Illuminate\Database\Eloquent\Model;
 use Filament\Tables\Columns\IconColumn;
-use App\Tables\Columns\CountDownColumn;
 use Illuminate\Database\Eloquent\Builder;
 use Filament\Forms\Components\Placeholder;
 use Filament\Infolists\Components\TextEntry;
@@ -24,6 +33,11 @@ use App\Filament\Resources\ProjectResource\Pages;
 use Illuminate\Database\Eloquent\SoftDeletingScope;
 use BezhanSalleh\FilamentShield\Contracts\HasShieldPermissions;
 use App\Filament\Resources\ProjectResource\RelationManagers\TestsRelationManager;
+use JaOcero\ActivityTimeline\Components\ActivityDate;
+use JaOcero\ActivityTimeline\Components\ActivityDescription;
+use JaOcero\ActivityTimeline\Components\ActivityIcon;
+use JaOcero\ActivityTimeline\Components\ActivitySection;
+use JaOcero\ActivityTimeline\Components\ActivityTitle;
 
 class ProjectResource extends Resource implements HasShieldPermissions
 {
@@ -58,6 +72,9 @@ class ProjectResource extends Resource implements HasShieldPermissions
             'set_done_project_test',
             'set_failed_project_test',
             'renewal_project_test',
+            'add_project_note',
+            'pause_project',
+            'continue_project',
         ];
     }
 
@@ -66,10 +83,12 @@ class ProjectResource extends Resource implements HasShieldPermissions
         return $infolist
             ->schema([
                 TextEntry::make('product.title')
-                    ->label('نام محصول'),
+                    ->formatStateUsing(fn (Model $record): string => $record->title ?? $record->product->title)
+                    ->label('نام'),
 
                 TextEntry::make('user.name')
                     ->label('نام کارمند'),
+
 
                 TextEntry::make('started_at')->label('شروع پروژه')->jalaliDate(),
 
@@ -83,26 +102,94 @@ class ProjectResource extends Resource implements HasShieldPermissions
                         fn (Model $record): string => verta(
                             $record->started_at
                                 ->addMinutes(
-                                    $record->tests->sum('duration') + $record->tests->sum('projectTest.renewals_duration')
+                                    $record->tests->sum('duration') + $record->tests->sum(
+                                        'projectTest.renewals_duration'
+                                    )
                                 )
                         )->format('H:i:s - Y/m/d')
                     ),
-
-
+                TextEntry::make('finished_at')->label('پایان یافته در')->jalaliDate(),
                 IconEntry::make('updated_at')
                     ->label('وضعیت')
                     ->icon(
-                        fn (Project $record): string => $record->finished_at
-                        ? ($record->is_mismatched ? 'heroicon-o-x-circle' : 'heroicon-o-check-circle')
-                        : 'heroicon-o-play-circle'
+                        function (Project $record): string {
+                            if ($record->isPaused()) {
+                                return 'heroicon-o-pause-circle';
+                            }
+                            return $record->finished_at
+                                 ? ($record->is_mismatched ? 'heroicon-o-x-circle' : 'heroicon-o-check-circle')
+                                 : 'heroicon-o-play-circle';
+                        }
                     )
                     ->color(
-                        fn (Project $record): string => $record->finished_at
-                        ? ($record->is_mismatched ? 'danger' : 'success')
-                        : 'info'
+                        function (Project $record): string {
+                            if ($record->isPaused()) {
+                                return 'warning';
+                            }
+                            return  $record->finished_at
+                                  ? ($record->is_mismatched ? 'danger' : 'success')
+                                  : 'info';
+                        }
                     ),
+                ActivitySection::make('activities')
+                    ->label('اقدامات')
+                    //->description('These are the activities that have been recorded.')
+                    ->schema([
+                        ActivityTitle::make('description')
+                            ->allowHtml(),
+                        // Be aware that you will need to ensure that the HTML is safe to render, otherwise your application will be vulnerable to XSS attacks.
+                        ActivityDate::make('created_at')
+                            ->getStateUsing(fn ($record) => verta($record->created_at)->format(' H:i:s Y/m/d '))
+                            ->date('Y/m/d H:i:s', 'Asia/Tehran'),
+                        ActivityIcon::make('status')
+                            ->icon(fn (string|null $state): string|null => match ($state) {
+                                'ideation' => 'heroicon-m-light-bulb',
+                                'drafting' => 'heroicon-m-bolt',
+                                'reviewing' => 'heroicon-m-document-magnifying-glass',
+                                'published' => 'heroicon-m-rocket-launch',
+                                default => 'heroicon-m-bolt',
+                            })
+                            ->color(fn (string|null $state): string|null => match ($state) {
+                                'ideation' => 'purple',
+                                'drafting' => 'info',
+                                'reviewing' => 'warning',
+                                'published' => 'success',
+                                default => 'info',
+                            }),
+                    ])
+                    ->columnSpanFull()
+                    ->showItemsCount(20)
+                    ->showItemsLabel('View Old')
+                    ->showItemsIcon('heroicon-m-chevron-down')
+                    ->showItemsColor('gray')
+                    ->headingVisible(),
+                ActivitySection::make('notes')
+                    ->label('نوت‌ها')
+                    ->schema([
+                            ActivityTitle::make('body')
+                                ->label('متن')
+                                ->allowHtml(),
+                            // Be aware that you will need to ensure that the HTML is safe to render, otherwise your application will be vulnerable to XSS attacks.
+                            ActivityDescription::make('attachment')
+                                ->allowHtml()
+                                ->getStateUsing(
+                                    fn (Note $record) => $record->attachment ? "<a href='" . \Storage::url($record->attachment) . "' style='color:rgb(192, 132, 252) ' target='_blank'>پیوست</a>" : ""
+                                )
 
-
+                    ->placeholder(""),
+                ActivityDate::make('created_at')
+                    ->getStateUsing(fn ($record) => verta($record->created_at)->format(' H:i:s Y/m/d '))
+                    ->date('Y/m/d H:i:s', 'Asia/Tehran'),
+                ActivityIcon::make('status')
+                    ->icon('heroicon-m-document')
+                    ->color('info'),
+            ])
+            ->columnSpanFull()
+            ->showItemsCount(20)
+            ->showItemsLabel('View Old')
+            ->showItemsIcon('heroicon-m-chevron-down')
+            ->showItemsColor('gray')
+            ->headingVisible()
             ])
             ->columns(3);
     }
@@ -129,19 +216,33 @@ class ProjectResource extends Resource implements HasShieldPermissions
                         ->fill())
                     ->required()->native(false),
 
+                TextInput::make('title')
+                    ->label('نام پروژه'),
+                Select::make('extra_time')
+                    ->options([
+                        10 => 10,
+                        15 => 15,
+                        20 => 20
+                    ])
+                    ->required()
+                    ->default(10)
+                    ->label('وقت اضافه')->native(false),
+
+
                 Grid::make(1)
                     ->schema(function (Get $get): array {
                         $product = Product::with('tests')->find($get('product_id'));
-                        return $product?->tests
-                            ->map(
-                                fn (Test $test, mixed $index) => Placeholder::make('employee_number')->label('')->content(
-                                    "مرحله " . ($index + 1) . ": $test->title ($test->duration دقیقه)"
+                        $duration = $product?->tests->sum('duration') ?? 0;
+                        $count = $product?->tests->count() ?? 0;
+                        return $product ? [
+                            Placeholder::make('employee_number')
+                                ->label('')->content(
+                                    "در کل $count آزمایش " . " در {$duration} دقیقه "
                                 )
-                            )->toArray()
-                            ?? [];
+                        ] : [];
                     })
                     ->key('dynamicTypeFields'),
-            ]);
+            ])->columns(2);
     }
 
     public static function table(Table $table): Table
@@ -151,7 +252,8 @@ class ProjectResource extends Resource implements HasShieldPermissions
             ->columns([
 
                 TextColumn::make('product.title')->searchable()
-                    ->label('نام محصول'),
+                    ->formatStateUsing(fn (Model $record): string => $record->title ?? $record->product->title)
+                    ->label('نام پروژه'),
 
                 TextColumn::make('user.name')->searchable()
                     ->label('نام کارمند'),
@@ -159,56 +261,109 @@ class ProjectResource extends Resource implements HasShieldPermissions
                 TextColumn::make('started_at')->label('شروع پروژه')->jalaliDate(),
 
                 TextColumn::make('id')->label('تمدید شده')
-                    ->formatStateUsing(fn (Model $record) => ($record->tests->sum('projectTest.renewals_duration')) . " دقیقه "),
+                    ->formatStateUsing(
+                        fn (Model $record) => ($record->tests->sum('projectTest.renewals_duration')) . " دقیقه "
+                    ),
 
                 TextColumn::make('product_id')->label('پایان تخمینی')
                     ->formatStateUsing(
                         fn (Model $record): string => verta(
                             $record->started_at
                                 ->addMinutes(
-                                    $record->tests->sum('duration') + $record->tests->sum('projectTest.renewals_duration')
+                                    $record->tests->sum('duration') + $record->tests->sum(
+                                        'projectTest.renewals_duration'
+                                    )
                                 )
                         )->format('H:i:s Y-m-d')
                     ),
 
+                TextColumn::make('notes_count')->label('نوت‌ها')->counts('notes')
+                    ->badge(),
 
-                CountDownColumn::make('user_id')->label('زمان باقی مانده')
+
+                NewCountDownColumn::make('user_id')
+                    ->label('زمان باقی مانده')
                     ->formatStateUsing(function (Project $record): ?int {
                         if ($record->isFinished() || $record->isExpired()) {
                             return null;
                         }
 
-                        return now()->diffInSeconds($record->getFinishesAt());
+                        return (int)now()->diffInSeconds($record->getFinishesAt());
                     }),
                 TextColumn::make('finished_at')->label('زمان پایان')->jalaliDate(),
                 IconColumn::make('updated_at')
                     ->label('وضعیت')
                     ->icon(
-                        fn (Project $record): string => $record->finished_at
-                        ? ($record->is_mismatched ? 'heroicon-o-x-circle' : 'heroicon-o-check-circle')
-                        : 'heroicon-o-play-circle'
+                        function (Project $record): string {
+                            if ($record->isPaused()) {
+                                return 'heroicon-o-pause-circle';
+                            }
+                            return $record->finished_at
+                                ? ($record->is_mismatched ? 'heroicon-o-x-circle' : 'heroicon-o-check-circle')
+                                : 'heroicon-o-play-circle';
+                        }
                     )
                     ->color(
-                        fn (Project $record): string => $record->finished_at
-                        ? ($record->is_mismatched ? 'danger' : 'success')
-                        : 'info'
+                        function (Project $record): string {
+                            if ($record->isPaused()) {
+                                return 'warning';
+                            }
+                            return  $record->finished_at
+                                ? ($record->is_mismatched ? 'danger' : 'success')
+                                : 'info';
+                        }
                     ),
             ])
             ->filters([
-//                Tables\Filters\TrashedFilter::make()->label('آرشیو')
-//                    ->placeholder('آرشیو نشده‌ها')
-//                    ->trueLabel('همه')
-//                    ->falseLabel('فقط آرشیو شده‌ها')
+                Filter::make('created_at')
+                    ->form([
+                        JalaliDatePicker::make('created_from')->label('از'),
+                        JalaliDatePicker::make('created_until')->label('تا'),
+                    ])
+                    ->query(function (Builder $query, array $data): Builder {
+                        return $query
+                            ->when(
+                                $data['created_from'],
+                                fn (Builder $query, $date): Builder => $query->whereDate('created_at', '>=', $date),
+                            )
+                            ->when(
+                                $data['created_until'],
+                                fn (Builder $query, $date): Builder => $query->whereDate('created_at', '<=', $date),
+                            );
+                    }),
+                  SelectFilter::make('product_id')
+                    ->searchable()
+                    ->multiple()
+                    ->preload()
+                    ->label('محصول')
+                    ->relationship('product', 'title')
+                    ->getSearchResultsUsing(
+                        fn (string $search) => Product::where('title', 'like', "%{$search}%")
+                            ->limit(50)
+                            ->pluck('title', 'id')
+                    ),
+                SelectFilter::make('user_id')
+                    ->searchable()
+                    ->multiple()
+                    ->preload()
+                    ->label('ایجاد کننده')
+                    ->relationship('user', 'name')
+                    ->getSearchResultsUsing(
+                        fn (string $search) => User::where('name', 'like', "%{$search}%")
+                            ->limit(50)
+                            ->pluck('name', 'id')
+                    ),
             ])
             ->actions([
+
 //                Tables\Actions\ViewAction::make(),
 //                Tables\Actions\EditAction::make(),
+
             ])
             ->bulkActions([
-//                Tables\Actions\BulkActionGroup::make([
-////                    Tables\Actions\DeleteBulkAction::make(),
-////                    Tables\Actions\ForceDeleteBulkAction::make(),
-////                    Tables\Actions\RestoreBulkAction::make(),
+//                BulkActionGroup::make([
+                    ContinueAction::make('bulk-continue'),
+                    PauseAction::make('bulk-pause')
 //                ]),
             ]);
         //  ->poll(60);
@@ -217,7 +372,8 @@ class ProjectResource extends Resource implements HasShieldPermissions
     public static function getRelations(): array
     {
         return [
-            TestsRelationManager::class
+            TestsRelationManager::class,
+//            NotesRelationManager::class
         ];
     }
 
