@@ -3,6 +3,7 @@
 namespace App\Filament\Resources\ProjectResource\Pages;
 
 use App\Filament\Resources\ProjectResource;
+use App\Filament\Resources\ProjectResource\Actions\AddNoteAction;
 use Filament\Actions;
 use Filament\Resources\Pages\ViewRecord;
 
@@ -13,7 +14,12 @@ class ViewProject extends ViewRecord
     protected function getHeaderActions(): array
     {
         return [
-            Actions\EditAction::make(),
+            //Actions\EditAction::make(),
+            AddNoteAction::make('add-note'),
+            ProjectResource\Actions\SetDoneAction::make('set-done'),
+            ProjectResource\Actions\SetFailedAction::make('set-failed'),
+            ProjectResource\Actions\PauseAction::make('set-paused'),
+            ProjectResource\Actions\ContinueAction::make('set-continued')
         ];
     }
 }
