@@ -2,24 +2,41 @@
 
 namespace App\Models;
 
+use App\Enums\ProjectStatusEnum;
 use Carbon\Carbon;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\SoftDeletes;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\MorphMany;
+use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
+use Spatie\Activitylog\ActivitylogServiceProvider;
+use Spatie\Activitylog\Exceptions\InvalidConfiguration;
+use Spatie\Activitylog\LogOptions;
+use Spatie\MediaLibrary\HasMedia;
+use Spatie\MediaLibrary\InteractsWithMedia;
 
-class Project extends Model
+class Project extends Model implements HasMedia
 {
+    use InteractsWithMedia;
     use HasFactory;
     use SoftDeletes;
+
 
     protected $casts = [
         'started_at' => 'datetime',
         'finished_at' => 'datetime',
+        'status' => ProjectStatusEnum::class,
     ];
 
     protected $guarded = ['id'];
+
+    public function getActivitylogOptions(): LogOptions
+    {
+        return LogOptions::defaults();
+    }
 
     public function user(): BelongsTo
     {
@@ -29,6 +46,11 @@ class Project extends Model
     public function product(): BelongsTo
     {
         return $this->belongsTo(Product::class);
+    }
+
+    public function notes(): MorphMany
+    {
+        return $this->morphMany(Note::class, 'notable');
     }
 
     public function tests(): BelongsToMany
@@ -44,17 +66,43 @@ class Project extends Model
                 'is_mismatched',
                 'renewals_count',
                 'renewals_duration',
+                'passed_duration',
             ])->orderByPivot('order');
+    }
+
+    /**
+     * @throws InvalidConfiguration
+     */
+    public function activities(): MorphMany
+    {
+        return $this->morphMany(ActivitylogServiceProvider::determineActivityModel(), 'subject');
     }
 
     public function getFinishesAt(): ?Carbon
     {
-        $tests = $this->tests->whereNull('projectTest.finished_at')->sortBy('id');
-        $first = $tests->first();
-        return $first->projectTest->started_at
-            ?->addMinutes(
+        if ($this->isPaused()) {
+            return null;
+        }
+
+        $tests = $this->tests
+            ->whereNull('projectTest.finished_at')
+            ->sortBy('id');
+        $first = $tests->shift();
+        if ($first === null || !$first->projectTest->started_at) {
+            return null;
+        }
+
+        return now()
+            ->addSeconds($first->projectTest->getRemainingSeconds())
+            ->addMinutes(
                 $tests->sum('duration') + ($tests->sum('projectTest.renewals_duration'))
             );
+    }
+
+    public function getRemainingMinutes(): ?int
+    {
+        $finishedAt = $this->getFinishesAt();
+        return (int)$finishedAt?->diffInMinutes(now());
     }
 
 
@@ -70,11 +118,121 @@ class Project extends Model
 
     public function isExpired(): bool
     {
-        return $this->started_at && $this->getFinishesAt()->isBefore(now());
+        return $this->started_at && $this->getFinishesAt()?->isBefore(now());
+    }
+    public function isPaused(): bool
+    {
+        return $this->status === ProjectStatusEnum::PAUSED;
     }
 
-    public function isMismatched(): bool
+
+    public function addNote(string $body, ?string $attachment): void
     {
-        return $this->is_mismatched;
+        $this->notes()->create([
+            'user_id' => Auth::id(),
+            'body' => $body,
+            'attachment' => $attachment,
+        ]);
+    }
+
+
+    private function determineStatus(bool $isMismatched): void
+    {
+        DB::transaction(function () use ($isMismatched) {
+            $this->update([
+                'finished_at' => now(),
+                'is_mismatched' => $isMismatched
+            ]);
+            $this
+                ->tests
+                ->whereNull('projectTest.finished_at')
+                ->sortBy('projectTest.order')->each(fn (Test $test) => $test->projectTest->update([
+                    'finished_at' => now(),
+                    'is_mismatched' => $isMismatched
+                ]));
+        });
+
+        activity()
+            ->event('finished')
+            ->useLog('projects')
+            ->performedOn($this)
+            ->causedBy(Auth::user())
+            ->log(
+                " آزمایش "
+                . "توسط "
+                . $this->user->name
+                . " با وضعیت "
+                . ($isMismatched ? "'نامنطبق'" : "'منطبق'")
+                . " تمام شد. "
+            );
+
+        redirect("/admin/projects/{$this->id}");
+    }
+
+    public function setDone(): void
+    {
+        $this->determineStatus(false);
+    }
+
+    public function setFailed(): void
+    {
+        $this->determineStatus(true);
+    }
+
+    public function pause(string $reason): void
+    {
+        $this->update(['status' => ProjectStatusEnum::PAUSED]);
+
+        $currentTest = $this->tests
+            ->where('projectTest.started_at', '!=', null)
+            ->whereNull('projectTest.finished_at')
+            ->first();
+
+        $currentTest?->projectTest
+            ->increment(
+                'passed_duration',
+                $currentTest->projectTest->getPassedSeconds()
+            );
+
+        activity()
+          ->event('paused')
+          ->useLog('projects')
+          ->performedOn($this)
+          ->causedBy(Auth::user())
+          ->log(
+              " آزمایش "
+              . "توسط "
+              . $this->user->name
+              . " متوقف شد و این علت برای آن ذکر شد: "
+              . $reason
+          );
+
+        redirect("/admin/projects/{$this->id}");
+    }
+    public function continue(): void
+    {
+        $currentTest = $this->tests
+            ->where('projectTest.started_at', '!=', null)
+            ->whereNull('projectTest.finished_at')
+            ->first();
+
+        $currentTest?->projectTest->update(['started_at' => now()]);
+
+        $this->update(['status' => ProjectStatusEnum::PROCESSING]);
+
+
+        activity()
+          ->event('continued')
+          ->useLog('projects')
+          ->performedOn($this)
+          ->causedBy(Auth::user())
+            ->log(
+                " آزمایش "
+                . "توسط "
+                . $this->user->name
+                . " ادامه داده شد. "
+            );
+
+        redirect("/admin/projects/{$this->id}");
     }
 }
