@@ -3,10 +3,8 @@
 namespace App\Providers;
 
 use App\Models\User;
-use Filament\Facades\Filament;
 use Laravel\Pulse\Facades\Pulse;
 use Illuminate\Support\ServiceProvider;
-use Filament\Navigation\NavigationGroup;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -24,7 +22,7 @@ class AppServiceProvider extends ServiceProvider
     public function boot(): void
     {
         Pulse::users(function ($ids) {
-            return User::findMany($ids)->map(fn($user) => [
+            return User::findMany($ids)->map(fn ($user) => [
                 'id' => $user->id,
                 'name' => $user->name,
                 'extra' => $user->email,
