@@ -20,13 +20,14 @@ return [
         'title' => "پروژه",
         'plural' => "پروژه‌ها",
         'group' => '',
-        'filters'=>[
-            'tabs'=>[
-                'all'=>'همه',
-                'testing'=>'درحال آزمایش',
-                'done'=>'آماده تحویل',
-                'failed'=>'عدم تطابق',
-                'archived'=>'آرشیو شده',
+        'filters' => [
+            'tabs' => [
+                'all' => 'همه',
+                'testing' => 'درحال آزمایش',
+                'done' => 'منطبق',
+                'failed' => 'نامنطبق',
+                'paused' => 'متوقف شده',
+                'archived' => 'آرشیو شده',
             ]
         ]
     ],
